+++ extensions/memory-core/src/session-search-visibility.ts
@@ -1,0 +1,62 @@
+import type { OpenClawConfig } from "openclaw/plugin-sdk/memory-core-host-runtime-core";
+import type { MemorySearchResult } from "openclaw/plugin-sdk/memory-core-host-runtime-files";
+import {
+  extractTranscriptStemFromSessionsMemoryHit,
+  resolveTranscriptStemToSessionKeys,
+} from "openclaw/plugin-sdk/session-transcript-hit";
+import {
+  createAgentToAgentPolicy,
+  createSessionVisibilityGuard,
+  resolveEffectiveSessionToolsVisibility,
+} from "openclaw/plugin-sdk/session-visibility";
+
+export async function filterMemorySearchHitsBySessionVisibility(params: {
+  cfg: OpenClawConfig;
+  agentId: string;
+  requesterSessionKey: string | undefined;
+  sandboxed: boolean;
+  hits: MemorySearchResult[];
+}): Promise<MemorySearchResult[]> {
+  const visibility = resolveEffectiveSessionToolsVisibility({
+    cfg: params.cfg,
+    sandboxed: params.sandboxed,
+  });
+  const a2aPolicy = createAgentToAgentPolicy(params.cfg);
+  const guard = params.requesterSessionKey
+    ? await createSessionVisibilityGuard({
+        action: "history",
+        requesterSessionKey: params.requesterSessionKey,
+        visibility,
+        a2aPolicy,
+      })
+    : null;
+
+  const next: MemorySearchResult[] = [];
+  for (const hit of params.hits) {
+    if (hit.source !== "sessions") {
+      next.push(hit);
+      continue;
+    }
+    if (!params.requesterSessionKey || !guard) {
+      continue;
+    }
+    const stem = extractTranscriptStemFromSessionsMemoryHit(hit.path);
+    if (!stem) {
+      continue;
+    }
+    const keys = resolveTranscriptStemToSessionKeys({
+      cfg: params.cfg,
+      agentId: params.agentId,
+      stem,
+    });
+    if (keys.length === 0) {
+      continue;
+    }
+    const allowed = keys.some((key) => guard.check(key).allowed);
+    if (!allowed) {
+      continue;
+    }
+    next.push(hit);
+  }
+  return next;
+}
