+++ src/config/sessions.ts
@@ -1,3 +1,4 @@
+export * from "./sessions/combined-store-gateway.js";
 export * from "./sessions/group.js";
 export * from "./sessions/artifacts.js";
 export * from "./sessions/metadata.js";
