+++ src/config/sessions/combined-store-gateway.ts
@@ -1,0 +1,98 @@
+import { resolveDefaultAgentId } from "../../agents/agent-scope.js";
+import {
+  canonicalizeSpawnedByForAgent,
+  resolveStoredSessionKeyForAgentStore,
+} from "../../gateway/session-store-key.js";
+import { normalizeAgentId } from "../../routing/session-key.js";
+import type { OpenClawConfig } from "../types.openclaw.js";
+import { resolveStorePath } from "./paths.js";
+import { loadSessionStore } from "./store-load.js";
+import { resolveAllAgentSessionStoreTargetsSync } from "./targets.js";
+import type { SessionEntry } from "./types.js";
+
+function isStorePathTemplate(store?: string): boolean {
+  return typeof store === "string" && store.includes("{agentId}");
+}
+
+function mergeSessionEntryIntoCombined(params: {
+  cfg: OpenClawConfig;
+  combined: Record<string, SessionEntry>;
+  entry: SessionEntry;
+  agentId: string;
+  canonicalKey: string;
+}) {
+  const { cfg, combined, entry, agentId, canonicalKey } = params;
+  const existing = combined[canonicalKey];
+
+  if (existing && (existing.updatedAt ?? 0) > (entry.updatedAt ?? 0)) {
+    combined[canonicalKey] = {
+      ...entry,
+      ...existing,
+      spawnedBy: canonicalizeSpawnedByForAgent(cfg, agentId, existing.spawnedBy ?? entry.spawnedBy),
+    };
+  } else {
+    combined[canonicalKey] = {
+      ...existing,
+      ...entry,
+      spawnedBy: canonicalizeSpawnedByForAgent(
+        cfg,
+        agentId,
+        entry.spawnedBy ?? existing?.spawnedBy,
+      ),
+    };
+  }
+}
+
+export function loadCombinedSessionStoreForGateway(cfg: OpenClawConfig): {
+  storePath: string;
+  store: Record<string, SessionEntry>;
+} {
+  const storeConfig = cfg.session?.store;
+  if (storeConfig && !isStorePathTemplate(storeConfig)) {
+    const storePath = resolveStorePath(storeConfig);
+    const defaultAgentId = normalizeAgentId(resolveDefaultAgentId(cfg));
+    const store = loadSessionStore(storePath);
+    const combined: Record<string, SessionEntry> = {};
+    for (const [key, entry] of Object.entries(store)) {
+      const canonicalKey = resolveStoredSessionKeyForAgentStore({
+        cfg,
+        agentId: defaultAgentId,
+        sessionKey: key,
+      });
+      mergeSessionEntryIntoCombined({
+        cfg,
+        combined,
+        entry,
+        agentId: defaultAgentId,
+        canonicalKey,
+      });
+    }
+    return { storePath, store: combined };
+  }
+
+  const targets = resolveAllAgentSessionStoreTargetsSync(cfg);
+  const combined: Record<string, SessionEntry> = {};
+  for (const target of targets) {
+    const agentId = target.agentId;
+    const storePath = target.storePath;
+    const store = loadSessionStore(storePath);
+    for (const [key, entry] of Object.entries(store)) {
+      const canonicalKey = resolveStoredSessionKeyForAgentStore({
+        cfg,
+        agentId,
+        sessionKey: key,
+      });
+      mergeSessionEntryIntoCombined({
+        cfg,
+        combined,
+        entry,
+        agentId,
+        canonicalKey,
+      });
+    }
+  }
+
+  const storePath =
+    typeof storeConfig === "string" && storeConfig.trim() ? storeConfig.trim() : "(multiple)";
+  return { storePath, store: combined };
+}
